New class to interface with dynesty. Make it easier to support more samplers.
